Move features into single package
